refactor(game): cut over high-scores view into package modules
