+++ src/crimson/game/high_scores_view/__init__.py
@@ -1,0 +1,3 @@
+from .view import HighScoresView
+
+__all__ = ["HighScoresView"]
